chore: update docs per review
