+++ docs/fiddles/features/web-bluetooth/main.js
@@ -1,10 +1,13 @@
-const {app, BrowserWindow} = require('electron')
+const {app, BrowserWindow, ipcMain} = require('electron')
 const path = require('path')
 
 function createWindow () {
   const mainWindow = new BrowserWindow({
     width: 800,
-    height: 600
+    height: 600,
+    webPreferences: {
+      preload: path.join(__dirname, 'preload.js')
+    }    
   })
 
   mainWindow.webContents.on('select-bluetooth-device', (event, deviceList, callback) => {
@@ -14,29 +17,16 @@ function createWindow () {
     } 
   })
 
+  // Listen for a message from the renderer to get the response for the Bluetooth pairing.
+  ipcMain.on('bluetooth-pairing-response', (event, response) => {
+    bluetoothPinCallback(response)
+  })
+
   mainWindow.webContents.session.setBluetoothPairingHandler((details, callback) => {
-    const response = {}
-  
-    switch (details.pairingKind) {
-      case 'confirm': {
-        response.confirmed = confirm(`Do you want to connect to device ${details.deviceId}`)
-        break
-      }
-      case 'confirmPin': {
-        response.confirmed = confirm(`Does the pin ${details.pin} match the pin displayed on device ${details.deviceId}?`)
-        break
-      }
-      case 'providePin': {
-        const pin = prompt(`Please provide a pin for ${details.deviceId}`)
-        if (pin) {
-          response.pin = pin
-          response.confirmed = true
-        } else {
-          response.confirmed = false
-        }
-      }
-    }
-    callback(response)
+
+    bluetoothPinCallback = callback
+    // Send a message to the renderer to prompt the user to confirm the pairing.
+    mainWindow.webContents.send('bluetooth-pairing-request', details)
   })  
 
   mainWindow.loadFile('index.html')
